requirements.txt
sqlalchemy
streamlit
psycopg2
pandas
authlib.integrations.requests_client